# Use lazy dual string/Buffer representation for trace IDs