--- a/README.md
+++ b/README.md
@@ -12,6 +12,7 @@ with Zipkin-compatible data model.
 
  1. `git submodule init update`
  2. `npm test`
+ 3. `make build-node`
 
   [ci-img]: https://travis-ci.org/uber/jaeger-client-node.svg?branch=master
   [cov-img]: https://coveralls.io/repos/github/uber/jaeger-client-node/badge.svg?branch=master