--- a/benchmarks/flame_graph_loop.js
+++ b/benchmarks/flame_graph_loop.js
@@ -19,17 +19,21 @@
 // THE SOFTWARE.
 
 var ConstSampler = require('../dist/src/samplers/const_sampler.js').default;
-var InMemoryReporter = require('../dist/src/reporters/in_memory_reporter.js').default;
+var RemoteReporter = require('../dist/src/reporters/remote_reporter.js').default;
+var NoopReporter = require('../dist/src/reporters/noop_reporter.js').default;
 var Tracer = require('../dist/src/tracer.js').default;
 var opentracing = require('opentracing');
+var argv = require('minimist')(process.argv.slice(2));
 
-function server_client_span_loop(tracer) {
+function server_client_span_loop(tracer, carrier) {
     for ( var i = 0; i <= 0; i--) {
         // deserialize parent context from carrier
-        var parentContext = tracer.extract(opentracing.FORMAT_HTTP_HEADERS, {});
+        var parentContext = tracer.extract(opentracing.FORMAT_HTTP_HEADERS, carrier);
 
         // create server span from parent context
-        var serverSpan = tracer.startSpan('server-span');
+        var serverSpan = tracer.startSpan('server-span', {
+            childOf: parentContext
+        });
 
         // create client span
         var clientSpan = tracer.startSpan('client-span', {
@@ -39,15 +43,42 @@ function server_client_span_loop(tracer) {
         tracer.inject(clientSpan.context(), opentracing.FORMAT_HTTP_HEADERS, {});
         // finish client span
         clientSpan.finish();
+
         // finish server span
         serverSpan.finish();
     }
 }
 
+console.dir(argv);
+
+var reporter = new NoopReporter();
+if (argv.r == "udp") {
+    console.log("Using Remote Reporter");
+    reporter = new RemoteReporter();
+} else {
+    console.log("Using Noop Reporter");
+}
+
+var sampler = new ConstSampler(false);
+var flag = "0";
+if (argv.s) {
+    console.log("Sampling is at 100%");
+    sampler = new ConstSampler(true);
+    flag = "1";
+} else {
+    console.log("Sampling is off");
+}
+
 var tracer = new Tracer(
     'flamegraph-tracer',
-    new InMemoryReporter(),
-    new ConstSampler(true)
+    reporter,
+    sampler
 );
 
-server_client_span_loop(tracer);
+var carrier = {
+    "uber-trace-id": "123:123:0:" + flag
+};
+
+console.dir(carrier);
+
+server_client_span_loop(tracer, carrier);