--- a/src/span_context.js
+++ b/src/span_context.js
@@ -190,6 +190,12 @@ export default class SpanContext {
             return null;
         }
 
+        // Note: Number type in JS cannot represent the full range of 64bit unsigned ints,
+        // so using parseInt() on strings representing 64bit hex numbers only returns
+        // an approximation of the actual value. Fortunately, we do not depend on parsing
+        // the IDs with parseInt(), we are only using it to validate that the string is
+        // a valid hex number (which is faster than doing it manually).  We cannot use
+        // Int64(numberValue).toBuffer() because it throws exceptions on bad strings.
         let approxTraceId = parseInt(headers[0], 16);
         let NaNDetected = (isNaN(approxTraceId, 16) || approxTraceId === 0) ||
                           isNaN(parseInt(headers[1], 16)) ||