According to the basic server-client loop `benchmarks/flame_graph_loop.js` flamegraph, to/from strings ops on the span context are among the most expensive portions.  This change introduces a dual representation of trace IDs as `string` and `Buffer`. Whichever is available first is stored, and the other is only generated as needed. For most (unsampled) spans the `fromString` operation produces string IDs, and the outbound `toString` operation can use them as is, without even converting to `Buffer` representation, which is more expensive.

## To/FromString Benchmarks

The micro-benchmarks below show almost 10 times improvement for to/from string methods.

### Before benchmark

```
$ node ./benchmarks/span_context.js

Beginning Span Context Benchmark...
  2 tests completed.

  SpanContext:fromString x   274,035 ops/sec ±1.03% (90 runs sampled)
  SpanContext:toString   x 1,649,536 ops/sec ±1.01% (93 runs sampled)
```

### After benchmark

```
$ node benchmarks/span_context.js

Beginning Span Context Benchmark...
  2 tests completed.

  SpanContext:fromString x  2,463,840 ops/sec ±1.33% (91 runs sampled)
  SpanContext:toString   x 12,618,966 ops/sec ±1.21% (94 runs sampled)
```

## Core Flows Benchmarks

The `core_flow` benchmark shows moderate to no improvement

### Before benchmark

$ node benchmarks/core_flows.js

Beginning Main Request Flows...
  Fundamental Request flow with const tracer and http carrier x 41,873 ops/sec ±5.00% (73 runs sampled)
  Fundamental Request flow with const tracer and text carrier x 60,004 ops/sec ±10.43% (76 runs sampled)
  Fundamental Request flow with probabilistic tracer and http carrier x 42,920 ops/sec ±12.71% (76 runs sampled)
  Fundamental Request flow with probabilistic tracer and text carrier x 51,835 ops/sec ±28.55% (74 runs sampled)

### After benchmark

$ node benchmarks/core_flows.js

Beginning Main Request Flows...
  Fundamental Request flow with const tracer and http carrier x 47,763 ops/sec ±4.87% (75 runs sampled)
  Fundamental Request flow with const tracer and text carrier x 64,508 ops/sec ±11.43% (72 runs sampled)
  Fundamental Request flow with probabilistic tracer and http carrier x 42,601 ops/sec ±21.58% (66 runs sampled)
  Fundamental Request flow with probabilistic tracer and text carrier x 52,418 ops/sec ±33.48% (54 runs sampled)


![image](https://cloud.githubusercontent.com/assets/3523016/19296943/0b4746a4-900e-11e6-9247-d3e515c8b19f.png)