@yurishkuro 2016-10-12
Also cleans up README, fixes #23